+++ src/semweb/Vocabulary.cpp
@@ -168,7 +168,7 @@ bool Vocabulary::isTaxonomicProperty(const std::string_view &iri) {
 
 // FIXME: the frequency of parent/child properties and classes should be managed too!
 
-void Vocabulary::setFrequency(const std::string_view &iri, uint32_t frequency) {
+void Vocabulary::setFrequency(const std::string_view &iri, uint64_t frequency) {
 	if (frequency == 0) {
 		frequency_.erase(iri);
 	} else {
@@ -195,7 +195,7 @@ void Vocabulary::increaseFrequency(const std::string_view &iri) {
 	}
 }
 
-uint32_t Vocabulary::frequency(const std::string_view &iri) const {
+uint64_t Vocabulary::frequency(const std::string_view &iri) const {
 	auto it = frequency_.find(iri);
 	if (it == frequency_.end()) return 0;
 	else return it->second;
